% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{aheatmap.biotmle}
\alias{aheatmap.biotmle}
\title{Heatmap for class \code{biotmle}}
\usage{
aheatmap.biotmle(x, ..., designMat, term = 2, FDRcutoff = 0.05, top = 25)
}
\arguments{
\item{x}{object of class \code{biotmle} as produced by an appropriate call to
\code{biomarkertmle}}

\item{...}{additional arguments passed \code{NMF::aheatmap} as necessary}

\item{designMat}{a design matrix providing the contrasts to be dispalyed in
the heatmap (as would be passed to \code{limma::lmFit}).}

\item{term}{numeric value indicating the column of the design matrix that
corresponds to the expression covariate of interest.}

\item{FDRcutoff}{cutoff to be used in controlling the False Discovery Rate}

\item{top}{number of identified biomarkers to plot in the heatmap}
}
\description{
Heatmap for class \code{biotmle}
}

