% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomarkerTMLEout}
\alias{biomarkerTMLEout}
\title{Results obtained from running the biomarkertmle on illumina2007 data}
\format{A \code{biotmle} object containing the results of running
        \code{biomarkertmle}.
\describe{
  These results are included here for the sake of making the vignettes build
  more quickly. The user will likely not benefit from using this data set.
}}
\usage{
biomarkerTMLEout
}
\value{
A \code{biotmle} object containing results from \code{biomarkertmle}.
}
\description{
Example results obtained from running the TMLE-based estimation procedure on
the example data included with this package (\code{illuminaData}).
}
\keyword{datasets}

