% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle.R
\docType{class}
\name{bioTMLE}
\alias{bioTMLE}
\alias{bioTMLE}
\alias{.biotmle}
\title{Constructor for class union data.frame_OR_EList}
\value{
fusion of classes \code{data.frame} and \code{EList}, used within
        \code{.biotmle} by class \code{bioTMLE} to handle uncertainty in the
        object passed to slot "tmleOut".

class \code{biotmle} object, sub-classed from SummarizedExperiment.
}
\description{
Constructor for class union data.frame_OR_EList

Constructor for class bioTMLE
}
\examples{
library(SummarizedExperiment)
library(biotmleData)
data(illuminaData)

example_biotmle_class <- function(se) {

    call <- match.call(expand.dots = TRUE)
    biotmle <- .biotmle(
          SummarizedExperiment(
             assays = assay(se),
             rowData = rowData(se),
             colData = colData(se)
          ),
          call = call,
          tmleOut = as.data.frame(matrix(NA, 10, 10)),
          topTable = as.data.frame(matrix(NA, 10, 10))
    )
    return(biotmle)
}

example_class <- example_biotmle_class(se = illuminaData)

}
