% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle.R
\name{biotmle}
\alias{biotmle}
\title{S3 constructor for class \code{biotmle}}
\usage{
biotmle(call = NULL, topTable = NULL, limmaOut = NULL, tmleOut = NULL)
}
\arguments{
\item{call}{merely passed in from \code{match.call(expand.dots = TRUE)}.}

\item{topTable}{\code{data.frame} containing the output generated by use of
\code{limma::topTable}.}

\item{limmaOut}{\code{MArrayList} containing the output generated by use of
\code{limma::lmFit}.}

\item{tmleOut}{\code{data.frame} containing the output generated by use of
\code{biomarkertmle}}
}
\value{
S3 object of class \code{biotmle} for TMLE-based biomarker analysis.
}
\description{
S3 constructor for class \code{biotmle}
}
