% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eif}
\alias{eif}
\title{Accessor for Table of Raw Efficient Influence Function Values}
\usage{
eif(object)
}
\arguments{
\item{object}{S4 object of class \code{bioTMLE}.}
}
\value{
contents of \code{tmleOut} slot of object of class \code{biotmle}.
}
\description{
Accessor for Table of Raw Efficient Influence Function Values
}
\examples{
library(dplyr)
library(biotmleData)
library(SuperLearner)
library(SummarizedExperiment)
data(illuminaData)

colData(illuminaData) <- colData(illuminaData) \%>\%
  data.frame() \%>\%
  mutate(age = as.numeric(age > median(age))) \%>\%
  DataFrame()
benz_idx <- which(names(colData(illuminaData)) \%in\% "benzene")

biomarkerTMLEout <- biomarkertmle(
  se = illuminaData[1:2, ],
  varInt = benz_idx,
  bppar_type = BiocParallel::SerialParam(),
  g_lib = c("SL.mean", "SL.glm"),
  Q_lib = c("SL.mean", "SL.glm")
)
eif(biomarkerTMLEout)
}
