% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{heatmap_biotmle}
\alias{heatmap_biotmle}
\title{Heatmap for class \code{biotmle}}
\usage{
heatmap_biotmle(biotmle)
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}
}
\description{
Heatmap for class \code{biotmle}
}

