% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illumina2007}
\alias{illumina2007}
\title{Baseline covariates and microarray data from 2007 study...}
\format{A data frame with 144 rows and 22196 variables:
\describe{
  ...some information ought to go here...
}}
\usage{
illumina2007
}
\description{
A dataset containing various baseline covariates and microarray expression
measures from Illumina arrays used in a 2007 study.
}
\keyword{datasets}

