% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{illuminaData}
\alias{illuminaData}
\title{Baseline covariates and Illumina microarray data from 2007 study}
\format{A \code{data.frame} with 144 rows and 22196 variables:
\describe{
  This is example data to be used in testing the \code{biomarkertmle}
  procedure. Consult the vignettes for how to use this data.
}}
\usage{
illuminaData
}
\value{
A \code{data.frame} containing biomarkers and baseline covariates.
}
\description{
A dataset containing various baseline covariates and microarray expression
measures from Illumina arrays used in a 2007 study.
}
\keyword{datasets}

