% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmatmle.R
\name{limmatmle}
\alias{limmatmle}
\title{The Moderated T-Statistic for Asymptotically Linear Parameters}
\usage{
limmatmle(biotmle, designMat, ...)
}
\arguments{
\item{biotmle}{\code{biotmle} object as generated by \code{biomarkertmle}}

\item{designMat}{a design matrix providing the contrasts to be used in the
linear model fitting procedure of \code{limma::lmFit}}

\item{...}{additional arguments to be passed to functions from \code{limma}}
}
\value{
\code{biotmle} object containing output from \code{limma::lmFit} and
        \code{limma::topTable}
}
\description{
Performs variance shrinkage via empirical Bayes procedures on influence
curve transforms of the Average Treatment Effect of gene expression
changes on the outcome of interest
}
\examples{
library(dplyr)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

W <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("age", "sex", "smoking"))) \%>\%
 dplyr::mutate(
   age = as.numeric((age > quantile(age, 0.25))),
   sex = I(sex),
   smoking = I(smoking)
 )

A <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("benzene")))
A <- A[, 1]

Y <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%ni\% c("age", "sex", "smoking", "benzene",
                                        "id")))
geneIDs <- colnames(Y)

design <- as.data.frame(cbind(rep(1, nrow(Y)),
                              as.numeric(A == max(unique(A)))))
colnames(design) <- c("intercept", "Tx")

limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
                          designMat = design)

}

