% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mt}
\alias{plot_mt}
\title{Plot p-values from moderated statistical tests for class biotmle}
\usage{
plot_mt(biotmle, type = "pvals_adj")
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}

\item{type}{character describing whether to provide a plot of unadjusted or
adjusted p-values (adjustment performed via Benjamini-Hochberg)}
}
\value{
object of class \code{ggplot} containing a histogram of the raw or
        Benjamini-Hochberg corrected p-values (depending on user input).
}
\description{
Histogram of raw or FDR-adjusted p-values from the moderated t-test.
}
\examples{
library(dplyr)
library(biotmleData)
library(SummarizedExperiment)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

colData(illuminaData) <- colData(illuminaData) \%>\%
     data.frame \%>\%
     dplyr::mutate(age = as.numeric(age > median(age))) \%>\%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
design <- as.numeric(designVar == max(designVar))

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)

plot_mt(biotmle = limmaTMLEout, type = "pvals_adj")

}
