% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetedLimma.R
\name{tmle.limma}
\alias{tmle.limma}
\title{Variance Shrinkage of Linear Models for Microarrays on ATE parameter
estimated via Targeted Minimum Loss-Based Estimation}
\usage{
tmle.limma(ATE, Y)
}
\arguments{
\item{ATE}{the average treatment effect of gene expression change on an
outcome of interest, as computed by \code{tmle.multiATE}}

\item{Y}{the vector corresponding to the outcome of interest}
}
\value{
A list containing two objects: (1) model fit returned by \code{limma::lmFit}
                               (2) results table from \code{limma::topTable}
}
\description{
Performs variance shrinkage via empirical Bayes procedures on influence
curve transforms of the Average Treatment Effect of gene expression
changes on the outcome of interest
}
\examples{

}

