% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformIC.R
\name{tmle.multi}
\alias{tmle.multi}
\title{The Influence Curve (IC) transform of genomic data via TMLE}
\usage{
tmle.multi(W, A, Y, a = 1, g.lib, Q.lib, family = "binomial",
  parallel = FALSE)
}
\arguments{
\item{W}{the data.frame corresponding to gene expression or similar data}

\item{A}{the covariate whose effect on the outcome is to be estimated}

\item{Y}{the outcome of interest, thought to be related to genes W}

\item{a}{the level of A against which comparisons performed (default = 1)}

\item{g.lib}{library...}

\item{Q.lib}{library...}

\item{family}{type of link function to be used given the outcome of interest}

\item{parallel}{whether to use foreach style parallelization procedure}
}
\value{
A list of data.frames for the multi-level TMLE estimates of genomic effects.
}
\description{
Computes the risk difference parameter for the expression level of a gene
on the outcome of interest via Targeted Minimum Loss-Based Estimation
}
\examples{

}

