% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{volcano_biotmle}
\alias{volcano_biotmle}
\title{Volcano plot for class biotmle}
\usage{
volcano_biotmle(biotmle)
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}
}
\value{
object of class \code{ggplot} containing a standard volcano plot of
        the log-fold change in the causal target parameter against the raw
        log p-value computed from the moderated t-test in \code{limmatmle}.
}
\description{
Volcano plot of the log-changes in the target causal paramter against the
log raw p-values from the moderated t-test.
}
\examples{
library(dplyr)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

W <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("age", "sex", "smoking"))) \%>\%
 dplyr::mutate(
   age = as.numeric((age > quantile(age, 0.25))),
   sex = I(sex),
   smoking = I(smoking)
 )

A <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("benzene")))
A <- A[, 1]

Y <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%ni\% c("age", "sex", "smoking", "benzene",
                                        "id")))
geneIDs <- colnames(Y)
design <- as.data.frame(cbind(rep(1, nrow(Y)),
                             as.numeric(A == max(unique(A)))))
colnames(design) <- c("intercept", "Tx")
limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
                          designMat = design)
volcano_biotmle(biotmle = limmaTMLEout)

}

