% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{volcanoPlot_biotmle}
\alias{volcanoPlot_biotmle}
\title{Volcano plot for class \code{biotmle}}
\usage{
volcanoPlot_biotmle(biotmle)
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}
}
\description{
Volcano plot for class \code{biotmle}
}

