% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_biomarkertmle-method.R
\name{biomarkerTMLE_exposure}
\alias{biomarkerTMLE_exposure}
\title{TMLE procedure using ATE for Biomarker Identication from Exposure}
\usage{
biomarkerTMLE_exposure(Y, W, A, a, subj_ids = NULL,
  family = "gaussian", g_lib, Q_lib)
}
\arguments{
\item{Y}{A \code{numeric} vector of expression values for a single biomarker.}

\item{W}{A \code{Matrix} of \code{numeric} values corresponding to baseline
covariates to be marginalized over in the estimation process.}

\item{A}{A \code{numeric} vector of discretized exposure vector (e.g., from a
design matrix whose effect on expression values is of interest.}

\item{a}{The \code{numeric} value indicating levels of \code{A} above against
which comparisons are to be made.}

\item{subj_ids}{A \code{numeric} vector of subject IDs to be passed directly
to \code{tmle::tmle} when there are repeated measures; measurements on the
same subject should have the exact same numerical identifier. These values
will be coerced to numeric if not provided in the appropriate form (e.g., as
\code{character}). The call to \code{tmle::tmle} will utilized a corrected
version of the variance estimate from the efficient influence function.}

\item{family}{(character) - specification of error family: "binomial" or
"gaussian"}

\item{g_lib}{(char vector) - library of learning algorithms to be used in
fitting the "g" step of the standard TMLE procedure.}

\item{Q_lib}{(char vector) - library of learning algorithms to be used in
fitting the "Q" step of the standard TMLE procedure.}
}
\value{
TMLE-based estimate of the relationship between biomarker expression
 and changes in an exposure variable, computed iteratively and saved in the
 \code{tmleOut} slot in a \code{biotmle} object.
}
\description{
This function performs influence curve-based estimation of the effect of an
exposure on biological expression values associated with a given biomarker,
controlling for a user-specified set of baseline covariates.
}
