% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerTMLE_outcome.R
\name{biomarkerTMLE_outcome}
\alias{biomarkerTMLE_outcome}
\title{TMLE procedure for Biomarker Identication from Outcome}
\usage{
biomarkerTMLE_outcome(Y, W, A, a, family = "gaussian", g_lib, Q_lib)
}
\arguments{
\item{Y}{(numeric vector) - a vector of array expression values for a single
gene}

\item{W}{(numeric matrix) - a matrix of covariates to be controlled in
estimation}

\item{A}{(numeric vector) - a discretized exposure vector whose effect on
expression gene expression values is of interest}

\item{a}{(numeric vector) - the levels of A against which comparisons are to
be made}

\item{family}{(character) - specification of error family: "binomial" or
"gaussian"}

\item{g_lib}{(char vector) - library of learning algorithms to be used in ...}

\item{Q_lib}{(char vector) - library of learning algorithms to be used in ...}
}
\description{
This function performs influence curve-based estimation of the effect of an
exposure on (gene) expression array values associated with a given biomarker,
controlling for a user-specified set of baseline covariates
}

