% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomarkerTMLEout}
\alias{biomarkerTMLEout}
\title{Results obtained from running the `biomarkertmle` on `illumina2007` data}
\format{A \code{biotmle} object containing the results of running
        \code{biomarkertmle}.
\describe{
  ...some information ought to go here...
}}
\usage{
biomarkerTMLEout
}
\description{
description...
}
\keyword{datasets}

