% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle_class.R
\docType{class}
\name{bioTMLE-class}
\alias{bioTMLE-class}
\alias{.biotmle}
\title{Constructor for class bioTMLE}
\value{
class \code{biotmle} object, sub-classed from SummarizedExperiment.
}
\description{
Constructor for class bioTMLE
}
\examples{
library(SummarizedExperiment)
library(biotmleData)
data(illuminaData)

example_biotmle_class <- function(se) {

    call <- match.call(expand.dots = TRUE)
    biotmle <- .biotmle(
          SummarizedExperiment(
             assays = assay(se),
             rowData = rowData(se),
             colData = colData(se)
          ),
          call = call,
          tmleOut = as.data.frame(matrix(NA, 10, 10)),
          topTable = as.data.frame(matrix(NA, 10, 10))
    )
    return(biotmle)
}

example_class <- example_biotmle_class(se = illuminaData)
}
