% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{data.frame_OR_EList-class}
\alias{data.frame_OR_EList-class}
\title{S4 class union data.frame_OR_EList}
\value{
fusion of classes \code{data.frame} and \code{EList}, used within
 \code{.biotmle} by class \code{bioTMLE} to handle uncertainty in the object
 passed to slot "tmleOut".
}
\description{
Virtual class union containing members of both \code{data.frame}
 and \code{limma::Elist}, used internally to handle situations when a
 returned object has a type that cannot be guessed from the function call.
}
