% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle-class.R
\name{eif}
\alias{eif}
\title{Accessor for Table of Raw Efficient Influence Function Values}
\usage{
eif(object)
}
\arguments{
\item{object}{S4 object of class \code{bioTMLE}.}
}
\description{
Accessor for Table of Raw Efficient Influence Function Values
}
