% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_ic}
\alias{heatmap_ic}
\title{Heatmap for class biotmle}
\usage{
heatmap_ic(x, ..., design, FDRcutoff = 0.05, top = 25)
}
\arguments{
\item{x}{object of class \code{biotmle} as produced by an appropriate call to
\code{biomarkertmle}}

\item{...}{additional arguments passed to \code{superheat::superheat} as
necessary}

\item{design}{a vector providing the contrast to be displayed in the heatmap.}

\item{FDRcutoff}{cutoff to be used in controlling the False Discovery Rate}

\item{top}{number of identified biomarkers to plot in the heatmap}
}
\value{
heatmap (from the superheat package) using hierarchical clustering to
        plot the changes in the variable importance measure for all subjects
        across a specified top number of biomarkers.
}
\description{
Heatmap of the contributions of a select subset of biomarkers to the variable
importance measure changes as assessed by influence curve-based estimation,
across all subjects.
}
\examples{
library(dplyr)
library(biotmleData)
library(SummarizedExperiment)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

colData(illuminaData) <- colData(illuminaData) \%>\%
     data.frame \%>\%
     dplyr::mutate(age = as.numeric(age > median(age))) \%>\%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
design <- as.numeric(designVar == max(designVar))

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
                           design = design)

heatmap_ic(x = limmaTMLEout, design = design, FDRcutoff = 0.05,
           top = 25)

}
