% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetedLimma.R
\name{limmaTMLE}
\alias{limmaTMLE}
\title{Moderated T-Statistic Test for Asymptotically Linear Target Parameters}
\usage{
limmaTMLE(biomarkerTMLEout, IDs = NULL, designMat, ...)
}
\arguments{
\item{biomarkerTMLEout}{a generic object generated as output from the TMLE
procedure implemented in \code{biomarkerTMLE}}

\item{IDs}{a vector providing the annotation information for the biomarkers,
genes, or other units of interested in the analysis}

\item{designMat}{a design matrix providing the contrasts to be used in the
linear model fitting procedure of \code{limma::lmFit}}

\item{...}{additional arguments to be passed to functions from \code{limma}}
}
\value{
A list containing two objects: (1) model fit returned by \code{limma::lmFit}
                               (2) results table from \code{limma::topTable}
}
\description{
Performs variance shrinkage via empirical Bayes procedures on influence
curve transforms of the Average Treatment Effect of gene expression
changes on the outcome of interest
}

