% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetedLimma.R
\name{limmatmle}
\alias{limmatmle}
\title{Testing via the Moderated T-Statistic for Asymptotically Linear Parameters}
\usage{
limmatmle(biotmle, designMat, ...)
}
\arguments{
\item{biotmle}{\code{biotmle} object as generated by \code{biomarkertmle}}

\item{designMat}{a design matrix providing the contrasts to be used in the
linear model fitting procedure of \code{limma::lmFit}}

\item{...}{additional arguments to be passed to functions from \code{limma}}
}
\value{
\code{biotmle} object containing output from \code{limma::lmFit} and
\code{limma::topTable}
}
\description{
Performs variance shrinkage via empirical Bayes procedures on influence
curve transforms of the Average Treatment Effect of gene expression
changes on the outcome of interest
}

