% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modtest_ic.R
\name{modtest_ic}
\alias{modtest_ic}
\title{Moderated Statistical Tests for Influence Curves}
\usage{
modtest_ic(biotmle, design, ...)
}
\arguments{
\item{biotmle}{\code{biotmle} object as generated by \code{biomarkertmle}}

\item{design}{a design matrix providing the contrasts to be used in the
linear model fitting procedure of \code{limma::lmFit}}

\item{...}{additional arguments to be passed to functions from \code{limma}}
}
\value{
\code{biotmle} object containing output from \code{limma::lmFit} and
        \code{limma::topTable}
}
\description{
Performs variance shrinkage via the empirical Bayes procedure of LIMMA on the
observed data after a transformation moving the data to influence curve
space, based on the average treatment effect parameter.
}
\examples{
library(dplyr)
library(biotmleData)
library(SummarizedExperiment)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

colData(illuminaData) <- colData(illuminaData) \%>\%
     data.frame \%>\%
     dplyr::mutate(age = as.numeric(age > median(age))) \%>\%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
design <- as.numeric(designVar == max(designVar))

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
                          design = design)

}
