% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{plot.biotmle}
\alias{plot.biotmle}
\title{Plot utility for class biotmle}
\usage{
\method{plot}{biotmle}(x, ..., type = "pvals_adj")
}
\arguments{
\item{x}{object of class \code{biotmle} as produced by an appropriate call to
\code{biomarkertmle}}

\item{...}{additional arguments passed \code{plot} as necessary}

\item{type}{character describing whether to provide a plot of unadjusted or
adjusted p-values (adjustment performed via Benjamini-Hochberg)}
}
\value{
object of class \code{ggplot} containing a histogram of the raw or
        Benjamini-Hochberg corrected p-values (depending on user input).
}
\description{
Histogram of raw or FDR-adjusted p-values from the moderated t-test.
}
\examples{
library(dplyr)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

W <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("age", "sex", "smoking"))) \%>\%
 dplyr::mutate(
   age = as.numeric((age > quantile(age, 0.25))),
   sex = I(sex),
   smoking = I(smoking)
 )

A <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%in\% c("benzene")))
A <- A[, 1]

Y <- illuminaData \%>\%
 dplyr::select(which(colnames(.) \%ni\% c("age", "sex", "smoking", "benzene",
                                        "id")))
geneIDs <- colnames(Y)

design <- as.data.frame(cbind(rep(1, nrow(Y)),
                             as.numeric(A == max(unique(A)))))
colnames(design) <- c("intercept", "Tx")
limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
                          designMat = design)
plot(x = limmaTMLEout, type = "pvals_adj")

}

