% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{plot.biotmle}
\alias{plot.biotmle}
\title{Plot utility for class \code{biotmle}}
\usage{
\method{plot}{biotmle}(x, ..., type = "pvals_adj")
}
\arguments{
\item{x}{object of class \code{biotmle} as produced by an appropriate call to
\code{biomarkertmle}}

\item{...}{additional arguments passed \code{plot} as necessary}

\item{type}{character describing whether to provide a plot of unadjusted or
adjusted p-values (adjustment performed via Benjamini-Hochberg)}
}
\description{
Plot utility for class \code{biotmle}
}

