% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_biotmle.R
\name{plot_biotmle}
\alias{plot_biotmle}
\title{Plot utility for class \code{biotmle}}
\usage{
plot_biotmle(biotmle, type)
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}
}
\description{
Plot utility for class \code{biotmle}
}

