% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rnaseq.R
\name{rnaseq_ic}
\alias{rnaseq_ic}
\title{Transformation utility for using "voom" with biomarker TMLE procedure}
\usage{
rnaseq_ic(biotmle, weights = TRUE, ...)
}
\arguments{
\item{biotmle}{(bioTMLE) - subclass of \code{SummarizedExperiment} containing
next-generation sequencing (NGS) count data in the "assays" slot.}

\item{weights}{(logical) - whether to return quality weights of samples in
the output object.}

\item{...}{- other arguments to be passed to functions \code{limma::voom} or
\code{limma::voomWithQualityWeights} as appropriate.}
}
\value{
\code{EList} object containing voom-transformed "expression" measures
 of count data (actually, the mean-variance trend) in the "E" slot, to be
 passed into the biomarker TMLE procedure.
}
\description{
This function prepares next-generation sequencing data (counts) for use with
the biomarker TMLE procedure by invoking the voom transform of \code{limma}.
}
