% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle-class.R
\name{toptable}
\alias{toptable}
\title{Accessor for Results Table of Moderated Influence Function Hypothesis Test}
\usage{
toptable(object)
}
\arguments{
\item{object}{S4 object of class \code{bioTMLE}.}
}
\description{
Accessor for Results Table of Moderated Influence Function Hypothesis Test
}
