% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_ic}
\alias{volcano_ic}
\title{Volcano plot for class biotmle}
\usage{
volcano_ic(biotmle)
}
\arguments{
\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
call to \code{biomarkertmle}}
}
\value{
object of class \code{ggplot} containing a standard volcano plot of
        the log-fold change in the causal target parameter against the raw
        log p-value computed from the moderated t-test in \code{limmatmle}.
}
\description{
Volcano plot of the log-changes in the target causal paramter against the
log raw p-values from the moderated t-test.
}
\examples{
library(dplyr)
library(biotmleData)
library(SummarizedExperiment)
data(illuminaData)
data(biomarkertmleOut)
"\%ni\%" = Negate("\%in\%")

colData(illuminaData) <- colData(illuminaData) \%>\%
     data.frame \%>\%
     dplyr::mutate(age = as.numeric(age > median(age))) \%>\%
     DataFrame

varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
design <- as.numeric(designVar == max(designVar))

limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
                           design = design)

volcano_ic(biotmle = limmaTMLEout)

}
